refactor: 文档杂务及部分测试简并
