+++ docs/dev/图像模板匹配示例.md
@@ -5,6 +5,7 @@
 > 本文演示 `find_feature` 系列函数的用法。
 > 每个示例都是**可直接放入 `src/tasks/onetime/` 并注册运行**的完整任务文件，
 > 风格与 `src/tasks/test/` 下的调试任务保持一致。
+> 相关示例：[OCR 操作示例](文字识别示例.md)（屏幕文字识别与点击）。
 
 ---
