+++ docs/dev/文字识别示例.md
@@ -5,6 +5,7 @@
 > 本文演示四个核心屏幕识别/点击函数的用法。
 > 每个示例都是**可直接放入 `src/tasks/onetime/` 并注册运行**的完整任务文件，
 > 风格与 `src/tasks/test/` 下的调试任务保持一致。
+> 相关示例：[find_feature 操作示例](图像模板匹配示例.md)（模板匹配与点击）。
 
 ---
