+++ docs/dev/账号唯一ID与多账户覆盖默认逻辑.md
@@ -2,7 +2,7 @@
 
 返回：[开发文档索引](../zh-CN/index.md) / [README](https://github.com/AliceJump/ok-end-field/blob/master/README.md)
 
-本文说明当前版本下，从第一次启动软件开始，账号身份与账号覆盖配置的默认行为；登录切换的执行流程（GUI → 存储 → Mixin → login_flow）见 [切换账户流程分析](切换账户流程分析.md)。
+本文说明当前版本下，从第一次启动软件开始，账号身份与账号覆盖配置的默认行为；登录切换的执行流程（GUI → 存储 → Mixin → login_flow）见 [切换账户流程分析](切换账户流程分析.md)。日常操作步骤见 [账号配置用户指南](../../zh-CN/账号配置用户指南.md)——本文只讲默认行为与实现逻辑，不重复配置步骤。
 
 ## 1. 首次启动时会发生什么
