+++ docs/zh-CN/账号配置用户指南.md
@@ -38,7 +38,7 @@
 5. 账号配置页看不到某个任务：
    - 该任务当前不支持多账号独立配置展示。
 
-## 6. 覆盖读取流程
+## 5. 覆盖读取流程
 
 ```mermaid
 flowchart TD
@@ -57,7 +57,7 @@ flowchart TD
     K -->|否| I
 ```
 
-## 7. 常见问题
+## 6. 常见问题
 
 1. 为什么账号配置页不填密码？
    - 因为切换账号依赖游戏登录页的「最近」账号列表，不输入密码。旧格式中的密码字段完全忽略。
@@ -68,7 +68,7 @@ flowchart TD
 4. 为什么建议填写完整手机号？
    - 程序按填写值的后四位点击「最近」账号。若多个最近账号后四位相同，可能选错账号；请确保后四位唯一。
 
-## 8. 相关文档
+## 7. 相关文档
 
 实现逻辑：[账号唯一ID与多账户覆盖默认逻辑](../dev/账号唯一ID与多账户覆盖默认逻辑.md)
