+++ tests/TestStateDrivenWaits.py
@@ -1,6 +1,8 @@
 import unittest
 from unittest.mock import patch
 
+from ok import Box
+
 from src.core.base_mixin.game_flow_mixin import GameFlowMixin
 from src.core.base_mixin.runtime_mixin import RuntimeMixin
 from src.tasks.mixin.battle_mixin import BattleMixin
@@ -302,6 +304,73 @@ class TestStateDrivenWaits(unittest.TestCase):
         self.assertEqual(task.keys, [("j", {})])
         self.assertEqual(task.click_kwargs, [{}])
         self.assertEqual(task.stable_waits, 1)
+
+    def test_in_team_falls_back_to_later_skill_template(self):
+        boxes = [Box(index * 100, 10, 20, 20) for index in range(1, 4)]
+
+        class StubTask:
+            _battle_member_count = 0
+
+            def _battle_feature_boxes(self, prefix):
+                return boxes
+
+            def find_one(self, feature, box):
+                matches = {
+                    ("skill_1", boxes[0].x): Box(101, 11, 10, 10, confidence=0.99),
+                    ("skill_2", boxes[1].x): None,
+                    ("skill_3", boxes[2].x): Box(301, 11, 10, 10, confidence=0.98),
+                }
+                return matches.get((feature, box.x))
+
+            def log_debug(self, message):
+                self.last_log = message
+
+        task = StubTask()
+        self.assertTrue(BattleMixin.in_team(task))
+        self.assertEqual(task._battle_member_count, 3)
+        self.assertIn("skill_3->框3", task.last_log)
+
+    def test_in_team_requires_two_skill_matches(self):
+        boxes = [Box(index * 100, 10, 20, 20) for index in range(1, 3)]
+
+        class StubTask:
+            _battle_member_count = 0
+
+            def _battle_feature_boxes(self, prefix):
+                return boxes
+
+            def find_one(self, feature, box):
+                if feature == "skill_1" and box.x == boxes[0].x:
+                    return Box(101, 11, 10, 10, confidence=0.99)
+                return None
+
+            def log_debug(self, message):
+                self.last_log = message
+
+        task = StubTask()
+        self.assertFalse(BattleMixin.in_team(task))
+        self.assertEqual(task._battle_member_count, 0)
+
+    def test_in_team_accepts_skill_one_in_last_slot(self):
+        boxes = [Box(index * 100, 10, 20, 20) for index in range(1, 5)]
+
+        class StubTask:
+            _battle_member_count = 0
+
+            def _battle_feature_boxes(self, prefix):
+                return boxes
+
+            def find_one(self, feature, box):
+                if feature == "skill_1" and box.x == boxes[-1].x:
+                    return Box(401, 11, 10, 10, confidence=0.99)
+                return None
+
+            def log_debug(self, message):
+                self.last_log = message
+
+        task = StubTask()
+        self.assertTrue(BattleMixin.in_team(task))
+        self.assertEqual(task._battle_member_count, 1)
 
 
 if __name__ == "__main__":
